net-next: mediatek: honour special tag bit inside RX DMA descriptor
For HW NAT/QoS to work the DSA driver needs to turn the special tag bit
inside the ingress control register on. This has the side effect that
the code working out which ingress gmac we have breaks. Fix this by
honouring the special tag bit inside the RX free descriptor.

Signed-off-by: John Crispin <john@phrozen.org>