s390/vfio-ap: check TAPQ response code when waiting for queue reset
The vfio_ap_mdev_reset_queue() function uses the PQAP(TAPQ) function to
verify that the reset completed by verifying that the queue is empty and
interrupts are disabled (a side effect of resetting the queue). If the
function can not verify the queue's status within 40ms, it logs a WARN
message to the system log indicating "unable to verify reset of queue
xx.yyyy".

The vfio_ap_mdev_reset_queue() function does not check the status
response code returned form the PQAP(TAPQ) function when verifying the
queue's status; consequently, there is no way of knowing whether
verification failed because the wait time was exceeded, or because the
PQAP(TAPQ) failed.

This patch adds code to check the status response code from the PQAP(TAPQ)
instruction and logs an appropriate message if it fails. It also improves
the current WARN message to include the queue status as well as the status
response code.

Reviewed-by: Matthew Rosato <mjrosato@linux.ibm.com>
Reviewed-by: Harald Freudenberger <freude@linux.ibm.com>
Signed-off-by: Tony Krowiak <akrowiak@linux.ibm.com>
Signed-off-by: Christian Borntraeger <borntraeger@linux.ibm.com>
Signed-off-by: Janosch Frank <frankja@linux.ibm.com>