i915/gvt: Move kvmgt_protect_table_init() and gvt_cache_init() into init
vfio_iommufd_bind() creates an access, which has an unmap callback, that
can be called immediately. So the dma_unmap() callback should tolerate the
unmaps that come before the emulated device is opened.

To achieve the above, move the protect_table_init() and gvt_cache_init()
into the init op which is supposed to be triggered prior to the
open_device() op.

Link: https://lore.kernel.org/r/20221123134832.429589-2-yi.l.liu@intel.com
Signed-off-by: Yi Liu <yi.l.liu@intel.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>