nvmem: add layout for Adtran devices
Adtran stores unique factory data on GPT partitions on the eMMC.
Using blk-nvmem the 'mfginfo' partition gets exposes as NVMEM provider.

Add layout driver to parse mfginfo, mainly to provide MAC addresses to
Ethernet and wireless interfaces.

Variable names are converted to lower-case and '_' is replaced with '-'
in order to comply with the device tree node naming convention.
The main MAC address always ends on a 0 and up to 16 addresses are
alocated for each device to use for various interfaces.

Implement post-processing function for 'MFG_MAC' variable ('mfg-mac'
node name in device tree) adding the nvmem cell index to the least
significant digit of the MAC address.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>