fs: compare truncated timestamps in current_mgtime
current_mgtime compares the ctime (which has already been truncated) to
the value from ktime_get_coarse_real_ts64 (which has not). All of the
existing filesystems that enable mgtime have 1ns granularity, so this is
not a problem today, but it is more correct to compare truncated
timestamps instead.

Do the truncate earlier, so we're comparing like things.

Signed-off-by: Jeff Layton <jlayton@kernel.org>
Message-Id: <20230728-mgctime-v1-1-5b0ddc5df08e@kernel.org>
Signed-off-by: Christian Brauner <brauner@kernel.org>