net: phylink: rtl822x SFP module no inband negotiation
The rtl822x does not seem to support inband negotiation, when it is in
"Rate Adaptor Mode". The datasheet has very limited information about
this mode.

Because there is no openly documented way to determine if the rtl822x is in
"Rate Adaptor Mode", assume when a rtl822x is usiing RollBall protocol
address, it is setup in this mode. Can change this, if a better option is
known to detect this mode.