KVM: x86: Use KVM-governed feature framework to track "XSAVES enabled"
Use the governed feature framework to track if XSAVES is "enabled", i.e.
if XSAVES can be used by the guest.  Add a comment in the SVM code to
explain the very unintuitive logic of deliberately NOT checking if XSAVES
is enumerated in the guest CPUID model.

No functional change intended.

Link: https://lore.kernel.org/r/20230729011608.1065019-13-seanjc@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>