LoongArch: Remove __noreturn attribute for die()
If notify_die() returns NOTIFY_STOP, there is no need to call the
make_task_dead() function. And then we can remove the '__noreturn'
attribute for die(), this is similar with arm64, riscv and csky.

While at it, modify the die() declaration in ptrace.h to fix the
following checkpatch warnings:

  WARNING: function definition argument 'const char *' should also have an identifier name
  WARNING: function definition argument 'struct pt_regs *' should also have an identifier name

Signed-off-by: Tiezhu Yang <yangtiezhu@loongson.cn>
Signed-off-by: Huacai Chen <chenhuacai@loongson.cn>