net: phy: realtek: Add driver instance for RTL8221-VB-CG via C45
Add driver instance for clause-45 communication with RTL8221-VB-CG. The driver
sets up a separate set of function pointers for clause-45 only accessable PHY's.
The C45 driver instance checks if the RTL8221-VB-CG is in, what Realtek
calls, "Rate Adaptor Mode".

Because there is no openly documented way to determine if the RTL8221-VB-CG
is in "Rate Adaptor Mode", assume when a RTL8221-VB-CG is on a sfp module and
using RollBall protocol address, it is setup in this mode.
Can change this, if a better option is known to detect this mode.

This "Rate Adaptor Mode" check is done inside the rtl822x_c45_xxx() functions.
At some point it could be changed at runtime, if data becomes available
on how this mode is switched on and off.

This new driver instance will be used by the OEM SFP-2.5G-T module. The
RTL8221-VB-CG registers are setup by the onboard microprocessor that is also
responsible for the RollBall protocol.

The C45 driver instance does not interfere with the settings applied to the
PHY by the microprocessor, if it is in "Rate Adaptor Mode". It only reads
information, using the RollBall protocol.