net: phy: realtek: rtlgen_get_speed(): Pass register value as argument
The register that needs to be read to determine the speed, is not
always read in the same manner, also not always at the same address.

To be able to use the one rtlgen_get_speed() function, generalize
rtlgen_get_speed() even more. Pass the value from the device specific
register as argument to rtlgen_get_speed().