ARM: dts: aspeed: Add ASRock E3C256D4I BMC
Like the E3C246D4I, this is a reasonably affordable off-the-shelf
mini-ITX AST2500/Xeon motherboard with good potential as an OpenBMC
development platform.  Booting the host requires a modicum of eSPI
support that's not yet in the mainline kernel, but most other basic
BMC functionality is available with this device-tree.

Signed-off-by: Zev Weiss <zev@bewilderbeest.net>
Link: https://lore.kernel.org/r/20231114112722.28506-6-zev@bewilderbeest.net
Signed-off-by: Joel Stanley <joel@jms.id.au>