mm: memcg: print out cgroup ino in the memcg tracepoints
Patch series "mm: memcg: improve vmscan tracepoints", v3.

The motivation behind this commit is to enhance the traceability and
understanding of memcg events.  By integrating the function cgroup_ino()
into the existing memcg tracepoints, this patch series introduces a new
tracepoint template for the begin() and end() events.  It utilizes a new
entry field ino to store the cgroup ino, enabling developers to easily
identify the cgroup associated with a specific memcg tracepoint event. 

Additionally, this patch series introduces new shrink_memcg tracepoints to
facilitate non-direct memcg reclaim tracing and debugging.  


This patch (of 2):

Sometimes, it becomes necessary to determine the memcg tracepoint event
that has occurred.  This is particularly relevant in scenarios involving a
large cgroup hierarchy, where users may wish to trace the process of
reclamation within a specific cgroup(s) by applying a filter.

The function cgroup_ino() is a useful tool for this purpose.  To integrate
cgroup_ino() into the existing memcg tracepoints, this patch introduces a
new tracepoint template for the begin() and end() events.

Link: https://lkml.kernel.org/r/20231123193937.11628-1-ddrokosov@salutedevices.com
Link: https://lkml.kernel.org/r/20231123193937.11628-2-ddrokosov@salutedevices.com
Signed-off-by: Dmitry Rokosov <ddrokosov@salutedevices.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Masami Hiramatsu (Google) <mhiramat@kernel.org>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Muchun Song <muchun.song@linux.dev>
Cc: Roman Gushchin <roman.gushchin@linux.dev>
Cc: Shakeel Butt <shakeelb@google.com>
Cc: Steven Rostedt (Google) <rostedt@goodmis.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>