mtd: ubi: block: make use of refcount_t
Use atomic operations provided for refcounts instead of relying in
the mutex protecting the block device' private data to also protect
the refcount.
Doing this allows checking if the device is busy without having to
acquire the mutex, which is beneficial as it allows to avoid
introducing complicated nested locking when creating ubi block devices
upon attachment triggered by notification.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>