net: phylink: add quirk for disabling in-band-status for mediatek pcs at 2500base-x
In follow up to: net: pcs: pcs-mtk-lynxi: use 2500Base-X without AN

MediaTek LynxI PCS, 2500Base-X will only work without inband status due to hardware
limitation.

Introduce bool phylink_major_no_inband(pl,interface), a function similar to
bool phylink_phy_no_inband(phy). Another option it could perhapse be done with a
bool pcs->ops->supports_inband(interface), where if the function-pointer is null,
it means it is supported for all.

Code added to phylink_major_config():

When there is no PHY attached, pl->pcs_neg_mode is set to PHYLINK_PCS_NEG_INBAND_DISABLED

When there is a PHY attached, pl->cur_link_an_mode is set to MLO_AN_PHY. To have the pcs
function correctly with the rtl8221b, we need to do the following to the in-band-status:

We need to disable it when interface of the pcs is set to 2500base-x, but need it
enable it when switched to sgmii.

So we get:

[ ... ] mtk_soc_eth 15100000.ethernet eth1: phy link up sgmii/1Gbps/Full/none/rx/tx
[ ... ] mtk_soc_eth 15100000.ethernet eth1: phylink_mac_config: mode=inband/sgmii/none
                                               adv=00,00000000,00000000,00000000 pause=03

[ ... ] mtk_soc_eth 15100000.ethernet eth1: phy link up 2500base-x/2.5Gbps/Full/none/rx/tx
[ ... ] mtk_soc_eth 15100000.ethernet eth1: phylink_mac_config: mode=phy/2500base-x/none
                                               adv=00,00000000,00008000,0000606f pause=03

 Changes to be committed:
	modified:   drivers/net/phy/phylink.c